conv_agent_vertex/requirements.txt
langchain
langchain-community
langchain-google-genai==2.1.4
firebase-admin
google-cloud-firestore
python-dotenv
flask
requests